requirements.txt
flask
python-freeipa
pyyaml
